Fix bug and ambiguity when using LLVM_SYS_<VERSION>_PREFIX
Before this change there was a bug where if you used
LLVM_SYS_<VERSION>_PREFIX, build.rs would only look for llvm-config
inside the prefix, and not the other binary candidates
llvm-config-X, llvm-config-X.Y and llvm-configXY.

Additionally with this change, if a prefix is specified, build.rs will
look for llvm-config ONLY inside the prefix. The assumption being that
the user know what they're doing, and may be trying to link to a
specific build or fork of LLVM.

If not prefix is set, $PATH is searched.